Events modal (#22)
# Description


https://github.com/orgs/GenerateNU/projects/31/views/1?pane=issue&itemId=53888866

Please include a summary of the changes and the related issue. Please
also
include relevant motivation, context, and images!

- Created a Event Page
- Modified Event Card to more closely resemble design mock-up
- Added Image Carosoule Functionality


![](https://files.slack.com/files-pri/T2CHL6FEG-F06LX54DQ8H/image_from_ios.png)

# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. If they
are unit
tests, provide the file name the tests are in. If they are not unit
tests,
describe how you tested the change.

# Checklist

- [ ] I have performed a self-review of my code
- [ ] I have reached out to another developer to review my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] New and existing unit tests pass locally with my changes

---------

Co-authored-by: ddusichka <78527291+ddusichka@users.noreply.github.com>