User handler tests (#43)
# Description

[Link to
Ticket](https://github.com/orgs/GenerateNU/projects/31/views/1?pane=issue&itemId=54810014)

Please include a summary of the changes and the related issue. Please
also
include relevant motivation, context, and images!

Tests for user operations at the handler level

# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. If they
are unit
tests, provide the file name the tests are in. If they are not unit
tests,
describe how you tested the change.

Ran tests locally, 30% code coverage for the handler package as a whole
and nearly 100% for the user handler specifically, except for a few
specific error cases.


![image](https://github.com/GenerateNU/couplet/assets/78527291/d99824bb-5094-400b-a7a0-bf35dff93bb1)

# Checklist

- [x] I have performed a self-review of my code
- [x] I have reached out to another developer to review my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] New and existing unit tests pass locally with my changes