added a list of matches to user struct and accounted for it in the us… (#45)
…er swipes endpoints

# Description

[Link to
Ticket](https://github.com/orgs/GenerateNU/projects/31/views/1?pane=issue&itemId=56341169)

Please include a summary of the changes and the related issue. Please
also
include relevant motivation, context, and images!

This MR adds a Matches field to the user struct so we can track a users
matches. When a user "swipes" another user, we'll determine if that
other user also liked the current user and if so, add them to each
other's matches list so we can easily display

# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. If they
are unit
tests, provide the file name the tests are in. If they are not unit
tests,
describe how you tested the change.

Created a new test for user swipes as I changed significant
functionality. I tested both scenarios as well to ensure we match users
accordingly. Can close other MR as that test is legacy now

# Checklist

- [ ] I have performed a self-review of my code
- [ ] I have reached out to another developer to review my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] New and existing unit tests pass locally with my changes