Event wireup (#50)
# Description

[Link to
Ticket](https://github.com/orgs/GenerateNU/projects/31/views/1?pane=issue&itemId=56332506)
@stoneliuCS 

Please include a summary of the changes and the related issue. Please
also
include relevant motivation, context, and images!

This PR wires up the home page and event cards to fetch events from the
backend. It uses the `GET("/events")` endpoint to fetch info for all
endpoints for the home page. When an event is clicked, the user is
routed to an Event page where the `GET("/events/{id}")` endpoint is used
to populate the page. The user can then react to events and these
changes will be saved to the DB.



https://github.com/GenerateNU/couplet/assets/78527291/db438dc2-509b-4e8a-84c3-1b11f4de6e7e


Some notes:
- needed to set baseUrl to local IP address when running locally for the
frontend to properly connect to backend
- noticed "id" was missing from the Event response type so I added it as
a (required) property

Future TODOs:
- [ ] match Figma more closely - we definitely have lots of room for
improvement with styling but wanted to wait for final designs and not
make this PR bigger than it already is
- [ ] Figure out idea of "collections" - how are we going to group
events into their rows on the home page?
- [ ] smoother loading and error display/handling - how should the user
know when something is wrong?
- [ ] some fields are missing from endpoint responses and/or db tables
that are needed on the frontend
- [ ] when fetching events, we should also fetch user reactions to them
(if there are any). we probably also want to prevent showing
already-reacted-to events in the card stack?

# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. If they
are unit
tests, provide the file name the tests are in. If they are not unit
tests,
describe how you tested the change.

# Checklist

- [ ] I have performed a self-review of my code
- [ ] I have reached out to another developer to review my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] New and existing unit tests pass locally with my changes

---------

Co-authored-by: Stone Liu <sliu@spillcenter.com>