Top Onboarding Bar + Screen ratio (#51)
# Description

N/A

Updated the top bar for onboarding to match the new hifis and added a
file to make using the screen width/height ratio easier.

# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. If they
are unit
tests, provide the file name the tests are in. If they are not unit
tests,
describe how you tested the change.

# Checklist

- [ ] I have performed a self-review of my code
- [ ] I have reached out to another developer to review my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] New and existing unit tests pass locally with my changes