Added share functionality (#79)
# Description

Please include a summary of the changes and the related issue. Please
also
include relevant motivation, context, and images!

Added share button functionality


# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. If they
are unit
tests, provide the file name the tests are in. If they are not unit
tests,
describe how you tested the change.

# Checklist

- [ ] I have performed a self-review of my code
- [ ] I have reached out to another developer to review my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] New and existing unit tests pass locally with my changes


https://github.com/GenerateNU/couplet/assets/23819935/9006739d-60da-40a9-8b26-bfc30634e69a