Settings tab (#87)
# Description

[Link to Ticket](insert the link to your ticket inside the parenthesis
here)

Please include a summary of the changes and the related issue. Please
also
include relevant motivation, context, and images!

This MR creates the Setting tab. Still need to hook up to backend.
(Honestly not sure how were accessing the current user right now) and
maybe some styling changes if needed



https://github.com/GenerateNU/couplet/assets/93676493/016f3f81-ab8f-4cae-bad3-293cf2c8e855



# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. If they
are unit
tests, provide the file name the tests are in. If they are not unit
tests,
describe how you tested the change.

# Checklist

- [ ] I have performed a self-review of my code
- [ ] I have reached out to another developer to review my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] New and existing unit tests pass locally with my changes