Update event schema (#90)
# Description

Please include a summary of the changes and the related issue. Please
also
include relevant motivation, context, and images!

This PR updates the event schema to include address, a minimum price, a
maximum price, and an external link. Only address and minPrice are
required. (MaxPrice is optional for admin convenience if there is only
one set price, and ExternalLink is optional in case a link simply
doesn't exist).

I also changed the type of Address from a custom object to a string for
now - this was a bit easier to work with, especially since the frontend
wants to display something concise

After this is merged, I can update the frontend to read from these
fields instead of using hardcoded values

# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. If they
are unit
tests, provide the file name the tests are in. If they are not unit
tests,
describe how you tested the change.

Ran tests locally and updated them with the new fields so they all
passed
Also ran data-gen a few times to make sure it populates the new fields

# Checklist

- [ ] I have performed a self-review of my code
- [ ] I have reached out to another developer to review my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] New and existing unit tests pass locally with my changes